Finish getting pipeline into an initially runnable state
Everything runs on a real data set from start to finish (minus the SVAs,
haven't ported those yet).

I had to replace some of the `.cpp` files with implementations that
aren't O(n²) to be able to run in a practical amount of time (and also
fixed a SIGSEGV in one).  I used CL just to get a proof-of-concept
working, I'll go back and rewrite them in bash or Python at some point.

This runs successfully on both Clover and Great Lakes pretty quickly
(15-20 minutes now, vs 41 hours for the same data set for the vanilla
version).