Add support for configurable sets of exclusions
The original Nanopal scripts have hardcoded in a couple of steps that
attempt to exclude non-reference-but-already-known MEIs from being
reported as novel MEIs, specifically MEIs that are already known to
exist in the GM12878 cell line.

That's fine when you're working with that cell line and want to know
what's new on top of it, but if you're not starting with that cell line
there's no particular reason to exclude those MEIs, e.g. for our HapMap
Mix data.  This adds support for disabling this extra exclusion step.