Add single-read-deduping, update PALMER
If a read has signal on both end, it gets two entries in some of the
intermediate files, one for each end.  This is normally fine, unless the
read is short enough that the two signals get clustered together into
a single hit.  Then that single reads gets counted twice in the read
support numbers, resulting in false positive multiple support numbers.

This change dedupes the `potential.clustered.txt.fi` before moving on,
so that any given read only counts as 1 read support.

I also had to remove the separate 5' and 3' counts from that file
because it's not clear which one I should subtract a duplicated read
from.  But as far as I know we've never used those numbers for anything
— they're dropped when we create the final BED file.